# [Pratt Parser] Optimize parser