WIP

master: 136830 ms to do 100 runs => 1368 ms/run

As of commit:
- 03d18c8 (head of pratt-parser): 186695 ms to do 100 runs => 1866.95 ms/run
- 883bacf (precise list of infix operators): 183531 ms to do 100 runs => 1835.31 ms/run
- 8ceeb9d (remove confusing off by one indent): 173865 ms to do 100 runs => 1738.65 ms/run
- c27fef3 (minus sign cannot be confused with a comment anymore):  166139 ms to do 100 runs => 1661.39 ms/run 
- cbbc2c6 (optimize multiline comment parser): 152409 ms to do 100 runs => 1524.09 ms/run