@jfmengels 2024-07-15
Thank you so much, this is a really nice performance increase!